+++ src/pygrits/viewpoints/__init__.py
@@ -1,1 +1,0 @@
-"""Example viewpoint schemas that extend the pygrits core."""